vip10

+++ app/Http/Controllers/UserController.php
@@ -280,17 +280,18 @@ class UserController extends Controller
 
        foreach($mainCourses as $mainCourse){
             $id=$mainCourse->course_id;
-          
-            if($req->$id=="on"){
-            DB::table('VipUsers')
-            ->updateOrInsert(
-                ['phone' => $phone, 'course_id' => $mainCourse->course_id],
-                ['major' => "$mainCourse->major",'course'=>"$mainCourse->title"]
-            );
+            if($mainCourse->major==$req->major){
+                if($req->$id=="on"){
+                DB::table('VipUsers')
+                ->updateOrInsert(
+                    ['phone' => $phone, 'course_id' => $mainCourse->course_id],
+                    ['major' => "$mainCourse->major",'course'=>"$mainCourse->title"]
+                );
         
-            }else{
-                VipUser::where('phone',$phone)->where('course_id',$mainCourse->course_id)->where('major',$mainCourse->major)->delete();
-            
+                }else{
+                    VipUser::where('phone',$phone)->where('course_id',$mainCourse->course_id)->where('major',$mainCourse->major)->delete();
+                
+                }
             }
             
         }
